Rename pages and verifiers to use their class names as file names
