+++ .eslintrc.js
@@ -11,6 +11,7 @@ module.exports = {
     'prettier',
     'prettier/@typescript-eslint',
     'plugin:prettier/recommended',
+    ''
   ],
   root: true,
   env: {
