+++ test/integration/registration.spec.ts
@@ -5,9 +5,9 @@ import {
   CREATE_USER_DTO_TEST_OBJ_WITH_INVALID_PASSWORD,
   CREATE_USER_DTO_TEST_OBJ_WITH_UNMATCHING_PASSWORDS
 } from '../../apps/backend/test/constants/users-test.constant';
-import RegistrationPage from '../support/pages/registration.page';
-import RegistrationVerifier from '../support/verifiers/registration.verifier';
-import ToastVerifier from '../support/verifiers/toast.verifier';
+import RegistrationPage from '../support/pages/RegistrationPage';
+import RegistrationVerifier from '../support/verifiers/RegistrationPageVerifier';
+import ToastVerifier from '../support/verifiers/ToastVerifier';
 
 context('Registration', () => {
   // Pages, verifiers, and modules
@@ -38,7 +38,7 @@ context('Registration', () => {
     it('rejects a weak password', () => {
       registrationPage.register(CREATE_USER_DTO_TEST_OBJ_WITH_INVALID_PASSWORD);
       toastVerifier.toastTextContains(
-        'Password does not meet complexity requirements. Passwords are a minimum of 15 characters in length. Passwords must contain at least one special character, number, upper-case letter, and lower-case letter. Passwords cannot contain more than three consecutive repeating characters. Passwords cannot contain more than four repeating characters from the same character class.'
+        'Password does not meet complexity requirements'
       );
     });
