+++ test/support/pages/RegistrationPage.ts
