+++ test/support/verifiers/RegistrationPageVerifier.ts
